+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - fix(core): use latest ubo-gui which adds bad-state-recovery for when menu-screen doesn't update - closes #235
 - chore(ci): rename `ubo-pod` runner label to `ubo-pod-pi4` for clarity
 - refactor(core): replace uuid4 with service path as the unique identifier of ubo services
+- refactor(core): introduce `utils.dataclass.default_provider` and use it wherever we had that duplicated code for setting the default value of a dataclass field based on other fields
 
 ## Version 1.3.0